+++ client/js/src/client.ts
@@ -12,7 +12,8 @@ import {
 	set_space_timeout,
 	hardware_types,
 	resolve_root,
-	apply_diff
+	apply_diff,
+	post_message
 } from "./utils.js";
 
 import type {
@@ -185,7 +186,8 @@ export function api_factory(
 	async function post_data(
 		url: string,
 		body: unknown,
-		token?: `hf_${string}`
+		token?: `hf_${string}`,
+		additional_headers?: Record<string, string>
 	): Promise<[PostResponse, number]> {
 		const headers: {
 			Authorization?: string;
@@ -199,7 +201,7 @@ export function api_factory(
 			var response = await fetch_implementation(url, {
 				method: "POST",
 				body: JSON.stringify(body),
-				headers
+				headers: { ...headers, ...additional_headers }
 			});
 		} catch (e) {
 			return [{ error: BROKEN_CONNECTION_MSG }, 500];
@@ -358,9 +360,10 @@ export function api_factory(
 
 				const _config = await config_success(config);
 				// connect to the heartbeat endpoint via GET request
-				const heartbeat = new EventSource(
+				const heartbeat_url = new URL(
 					`${config.root}/heartbeat/${session_hash}`
 				);
+				EventSource_factory(heartbeat_url); // Just connect to the endpoint without parsing the response. Ref: https://github.com/gradio-app/gradio/pull/7974#discussion_r1557717540
 				res(_config);
 			} catch (e) {
 				console.error(e);
@@ -438,15 +441,18 @@ export function api_factory(
 			): SubmitReturn {
 				let fn_index: number;
 				let api_info;
+				let dependency;
 
 				if (typeof endpoint === "number") {
 					fn_index = endpoint;
 					api_info = api.unnamed_endpoints[fn_index];
+					dependency = config.dependencies[endpoint];
 				} else {
 					const trimmed_endpoint = endpoint.replace(/^\//, "");
 
 					fn_index = api_map[trimmed_endpoint];
 					api_info = api.named_endpoints[endpoint.trim()];
+					dependency = config.dependencies[api_map[trimmed_endpoint]];
 				}
 
 				if (typeof fn_index !== "number") {
@@ -776,15 +782,27 @@ export function api_factory(
 								fn_index,
 								time: new Date()
 							});
-
-							post_data(
-								`${config.root}/queue/join?${url_params}`,
-								{
-									...payload,
-									session_hash
-								},
-								hf_token
-							).then(([response, status]) => {
+							let hostname = window.location.hostname;
+							let hfhubdev = "dev.spaces.huggingface.tech";
+							const origin = hostname.includes(".dev.")
+								? `https://moon-${hostname.split(".")[1]}.${hfhubdev}`
+								: `https://huggingface.co`;
+							const zerogpu_auth_promise =
+								dependency.zerogpu && window.parent != window && config.space_id
+									? post_message<Headers>("zerogpu-headers", origin)
+									: Promise.resolve(null);
+							const post_data_promise = zerogpu_auth_promise.then((headers) => {
+								return post_data(
+									`${config.root}/queue/join?${url_params}`,
+									{
+										...payload,
+										session_hash
+									},
+									hf_token,
+									headers
+								);
+							});
+							post_data_promise.then(([response, status]) => {
 								if (status === 503) {
 									fire_event({
 										type: "status",
@@ -1049,6 +1067,10 @@ export function api_factory(
 				event_stream = EventSource_factory(url);
 				event_stream.onmessage = async function (event) {
 					let _data = JSON.parse(event.data);
+					if (_data.msg === "close_stream") {
+						close_stream();
+						return;
+					}
 					const event_id = _data.event_id;
 					if (!event_id) {
 						await Promise.all(
@@ -1073,9 +1095,6 @@ export function api_factory(
 							pending_stream_messages[event_id] = [];
 						}
 						pending_stream_messages[event_id].push(_data);
-					}
-					if (_data.msg === "close_stream") {
-						close_stream();
 					}
 				};
 				event_stream.onerror = async function (event) {