+++ js/_spaces-test/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Dependency updates
 
+- @gradio/theme@0.2.1
+- @gradio/client@0.15.1
+- @gradio/form@0.1.13
+
+## 0.0.1
+
+### Dependency updates
+
 - @gradio/client@0.15.0
 
 ## 0.0.1